add serialize json plugin
