+++ packages/lix-sdk/src/snapshot/create-snapshot.ts
@@ -1,4 +1,3 @@
-import { sql } from "kysely";
 import type { Snapshot } from "../database/schema.js";
 import type { Lix } from "../lix/open-lix.js";
 
@@ -29,7 +28,6 @@ export async function createSnapshot(args: {
 				})),
 			)
 			.returningAll()
-			.returning(sql`json("content")`.as("content"))
 			.executeTakeFirstOrThrow();
 		return snapshot;
 	};
